requirements.txt
flask==3.0.0
pdfplumber==0.10.3
python-docx==1.1.0
spacy==3.7.2
keybert==0.8.1
sentence-transformers==2.2.2
pydantic==2.5.0
requests==2.31.0
pytest==7.4.3
